[ST-Engine] Add PipelineBaseScheduler For Engine (#2396)
* Add SeaTunnel Engine ConfigProvider and seatunnel-seatunnel-starter

* add source file to licenserc.yaml

* fix checkstyle

* fix error

* tmp

* tmp

* Update PhysicalPlan to support scheduler by pipeline

* remove init in run

* tmp

* complete scheduler

* optimize scheduler

* fix checkstyle

* fix checkstyle

* Update seatunnel-core/seatunnel-seatunnel-starter/src/main/java/org/apache/seatunnel/core/starter/seatunnel/SeaTunnelStarter.java


* Update SeaTunnelStarter.java

+++ seatunnel-engine/seatunnel-engine-core/src/main/java/org/apache/seatunnel/engine/core/dag/actions/PhysicalSourceAction.java
@@ -33,7 +33,7 @@ public class PhysicalSourceAction<T, SplitT extends SourceSplit, StateT extends 
     private final SeaTunnelSource<T, SplitT, StateT> source;
     private final List<SeaTunnelTransform> transforms;
 
-    public PhysicalSourceAction(int id,
+    public PhysicalSourceAction(long id,
                                 @NonNull String name,
                                 @NonNull SeaTunnelSource<T, SplitT, StateT> source,
                                 @NonNull List<URL> jarUrls,
@@ -43,7 +43,7 @@ public class PhysicalSourceAction<T, SplitT extends SourceSplit, StateT extends 
         this.transforms = transforms;
     }
 
-    protected PhysicalSourceAction(int id, @NonNull String name, @NonNull List<Action> upstreams,
+    protected PhysicalSourceAction(long id, @NonNull String name, @NonNull List<Action> upstreams,
                                    @NonNull SeaTunnelSource<T, SplitT, StateT> source,
                                    @NonNull List<URL> jarUrls,
                                    List<SeaTunnelTransform> transforms) {
